Moved Test Files to new Package "SearchFunctionalty"
