+++ src/main/java/com/xeeva/catalog/SearchItems/BPOItemsTab.java
@@ -2,13 +2,17 @@ package com.xeeva.catalog.SearchItems;
 
 import java.util.List;
 import java.util.ResourceBundle;
+
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
 import com.orasi.core.interfaces.Link;
 import com.orasi.core.interfaces.impl.internal.ElementFactory;
 import com.orasi.utils.Constants;
 import com.orasi.utils.OrasiDriver;
+import com.orasi.utils.PageLoaded;
 import com.orasi.utils.Sleeper;
+import com.orasi.utils.TestReporter;
 
 /**
  * @summary This page contains BPO Items Tab objects
@@ -16,6 +20,7 @@ import com.orasi.utils.Sleeper;
  * @date    06/09/16
  */
 public class BPOItemsTab {
+	PageLoaded pl = new PageLoaded();
 	private OrasiDriver driver = null;
 	private ResourceBundle userCredentialRepo = ResourceBundle.getBundle(Constants.USER_CREDENTIALS_PATH);
 
@@ -45,8 +50,9 @@ public class BPOItemsTab {
 	 **/
 	public void click_bpoItemsTab(){
 		pageLoaded();
-		bpoItemsTab.click();
-		Sleeper.sleep(3000);
+		//bpoItemsTab.click();
+		driver.executeJavaScript("arguments[0].click();", bpoItemsTab);
+		pl.isDomComplete(driver);
 	}
 
 	/**
@@ -58,12 +64,14 @@ public class BPOItemsTab {
 	public String  getBPOItemNumber(){
 		String ItemNumber = null;
 		click_bpoItemsTab();
-		List<WebElement> localItems= bpoItemsGrid;
-		if(localItems.size()>0){
-			for(WebElement inputItem :localItems){
+		List<WebElement> bpoItems= bpoItemsGrid;
+		if(bpoItems.size()>0){
+			for(WebElement inputItem :bpoItems){
 				ItemNumber = inputItem.getText();
 				break;
 			}
+		}else{
+			TestReporter.assertFalse(!(bpoItems.size()>0), "Global Items Count Empty!!");
 		}
 		return ItemNumber;
 	}
