+++ src/test/java/catalog/SearchFunctionality/CreateSmartFormRequest.java
@@ -1,4 +1,4 @@
-package catalog;
+package catalog.SearchFunctionality;
 
 import org.testng.ITestContext;
 import org.testng.annotations.AfterTest;
