+++ src/test/resources/catalog/SearchFunctionality/CreateSmartFormRequest.xml
@@ -6,9 +6,9 @@
 	<parameter name="runLocation" value="local" />
 	<parameter name="browserVersion" value="" />
 	<parameter name="operatingSystem" value="windows" />
- 	<test name="VerifyCatalogSearch">
+ 	<test name="CreateSmartFormRequest">
 		<classes>
-			<class name="catalog.VerifyCatalogSearch" />
+			<class name="catalog.SearchFunctionality.CreateSmartFormRequest" />
 		</classes>
 	</test> 
 </suite> <!-- Suite -->
