+++ src/test/resources/catalog/SearchFunctionality/VerifyCatalogSearch.xml
@@ -6,9 +6,9 @@
 	<parameter name="runLocation" value="local" />
 	<parameter name="browserVersion" value="" />
 	<parameter name="operatingSystem" value="windows" />
- 	<test name="CreateSmartFormRequest">
+ 	<test name="VerifyCatalogSearch">
 		<classes>
-			<class name="catalog.CreateSmartFormRequest" />
+			<class name="catalog.SearchFunctionality.VerifyCatalogSearch" />
 		</classes>
 	</test> 
 </suite> <!-- Suite -->
